+++ lib/que/scheduler/scheduler_job.rb
@@ -4,7 +4,7 @@ require_relative 'schedule'
 require_relative 'enqueueing_calculator'
 require_relative 'scheduler_job_args'
 require_relative 'state_checks'
-require_relative 'job_type_support'
+require_relative 'to_enqueue'
 require_relative 'version_support'
 
 # The main job that runs every minute, determining what needs to be enqueued, enqueues the required
@@ -23,13 +23,11 @@ module Que
 
           scheduler_job_args = SchedulerJobArgs.build(options)
           logs = ["que-scheduler last ran at #{scheduler_job_args.last_run_time}."]
-
           result = EnqueueingCalculator.parse(Scheduler.schedule.values, scheduler_job_args)
           enqueued_jobs = enqueue_required_jobs(result, logs)
           enqueue_self_again(
             scheduler_job_args, scheduler_job_args.as_time, result.job_dictionary, enqueued_jobs
           )
-
           # Only now we're sure nothing errored, log the results
           logs.each { |str| ::Que.log(event: 'que-scheduler'.to_sym, message: str) }
           destroy
@@ -38,23 +36,23 @@ module Que
 
       def enqueue_required_jobs(result, logs)
         result.missed_jobs.map do |to_enqueue|
-          enqueued_job = JobTypeSupport.enqueue(to_enqueue)
-          check_enqueued_job(enqueued_job, to_enqueue.job_class, to_enqueue.args, logs)
+          to_enqueue.enqueue.tap do |enqueued_job|
+            check_enqueued_job(to_enqueue, enqueued_job, logs)
+          end
         end.compact
       end
 
       private
 
-      def check_enqueued_job(enqueued_job, job_class, args, logs)
-        unless JobTypeSupport.valid_job_class?(enqueued_job.class)
-          # This can happen if a middleware nixes the enqueue call
-          logs << "que-scheduler called enqueue on #{job_class} but did not receive a #{Que::Job}"
-          return nil
-        end
-
-        job_id = JobTypeSupport.job_id(enqueued_job)
-        logs << "que-scheduler enqueueing #{job_class} #{job_id} with args: #{args}"
-        enqueued_job
+      def check_enqueued_job(to_enqueue, enqueued_job, logs)
+        logs << if enqueued_job.present?
+                  "que-scheduler enqueueing #{enqueued_job.job_class} " \
+                              "#{enqueued_job.job_id} with args: #{enqueued_job.args}"
+                else
+                  # This can happen if a middleware nixes the enqueue call
+                  "que-scheduler called enqueue on #{to_enqueue.job_class} " \
+                              'but it reported no job was scheduled. Has `enqueue` been overridden?'
+                end
       end
 
       def enqueue_self_again(scheduler_job_args, last_full_execution, job_dictionary, enqueued_jobs)
@@ -72,7 +70,7 @@ module Que
         )
 
         # rubocop:disable Style/GuardClause This reads better as a conditional
-        unless check_enqueued_job(enqueued_job, SchedulerJob, {}, [])
+        unless Que::Scheduler::VersionSupport.job_attributes(enqueued_job).fetch(:job_id)
           raise 'SchedulerJob could not self-schedule. Has `.enqueue` been monkey patched?'
         end
         # rubocop:enable Style/GuardClause