+++ spec/que/scheduler/audit_spec.rb
@@ -1,6 +1,8 @@
 require 'spec_helper'
 
 RSpec.describe Que::Scheduler::Audit do
+  include_context 'job testing'
+
   describe '.append' do
     def append_test_jobs(enqueued, executed_at, job_id)
       described_class.append(job_id, executed_at, enqueued)
@@ -20,41 +22,52 @@ RSpec.describe Que::Scheduler::Audit do
     it 'appends an audit line' do
       Timecop.freeze do
         scheduler_job_id = 1234
-        executed_at = Time.zone.now.change(usec: 0)
-        enqueued = [
-          HalfHourlyTestJob.enqueue(5, queue: 'something', run_at: executed_at - 1.hour),
-          HalfHourlyTestJob.enqueue(priority: 80, run_at: executed_at - 2.hours),
-          DailyTestJob.enqueue(3, queue: 'some_queue', run_at: executed_at - 3.hours),
+        # Roundtripping through postgres can yield usec differences which give spurious
+        # spec failures, so we ignore usec.
+        audit_insertion_time = Time.zone.now.change(usec: 0)
+        jobs_set_to_run_at = Que::Scheduler::Db.now.change(usec: 0)
+
+        te = Que::Scheduler::ToEnqueue
+        to_enqueue = [
+          te.create(job_class: HalfHourlyTestJob, args: 5, queue: 'something1'),
+          te.create(job_class: HalfHourlyTestJob, priority: 80),
+          te.create(job_class: DailyTestJob, args: 3, queue: 'something3', priority: 42),
         ]
-        db_jobs = append_test_jobs(enqueued, executed_at, scheduler_job_id)
+
+        enqueued = to_enqueue.map(&:enqueue)
+        # Do the other part of the usec removal from above here
+        db_jobs = append_test_jobs(enqueued, audit_insertion_time, scheduler_job_id).each { |row|
+          row[:run_at] = row[:run_at].change(usec: 0)
+        }
+
         expect(db_jobs).to eq(
           [
             {
               scheduler_job_id: scheduler_job_id,
               job_class: 'HalfHourlyTestJob',
-              queue: 'something',
+              queue: handles_queue_name ? 'something1' : nil,
               priority: 100,
               args: [5],
-              job_id: Que::Scheduler::JobTypeSupport.job_id(enqueued[0]),
-              run_at: executed_at - 1.hour,
+              job_id: enqueued[0].job_id,
+              run_at: jobs_set_to_run_at,
             },
             {
               scheduler_job_id: scheduler_job_id,
               job_class: 'HalfHourlyTestJob',
-              queue: Que::Scheduler.configuration.que_scheduler_queue,
+              queue: handles_queue_name ? Que::Scheduler.configuration.que_scheduler_queue : nil,
               priority: 80,
               args: [],
-              job_id: Que::Scheduler::JobTypeSupport.job_id(enqueued[1]),
-              run_at: executed_at - 2.hours,
+              job_id: enqueued[1].job_id,
+              run_at: jobs_set_to_run_at,
             },
             {
               scheduler_job_id: scheduler_job_id,
               job_class: 'DailyTestJob',
-              queue: 'some_queue',
-              priority: 100,
+              queue: handles_queue_name ? 'something3' : nil,
+              priority: 42,
               args: [3],
-              job_id: Que::Scheduler::JobTypeSupport.job_id(enqueued[2]),
-              run_at: executed_at - 3.hours,
+              job_id: enqueued[2].job_id,
+              run_at: jobs_set_to_run_at,
             },
           ]
         )