+++ spec/support/hash_support.rb
@@ -1,7 +1,7 @@
 module HashSupport
   class << self
     def hash_to_enqueues(overdue_dictionary)
-      overdue_dictionary.map { |item| Que::Scheduler::DefinedJob::ToEnqueue.new(item) }
+      overdue_dictionary.map { |item| Que::Scheduler::ToEnqueue.create(item) }
     end
   end
 end