+++ spec/support/test_jobs.rb
@@ -1,27 +1,28 @@
 require 'que'
 
-class HalfHourlyTestJob < ::Que::Job
-  def run; end
-end
+%w[
+  HalfHourlyTestJob
+  SpecifiedByClassTestJob
+  WithArgsTestJob
+  DailyTestJob
+  TwiceDailyTestJob
+  TimezoneTestJob
+].each do |name|
+  clazz =
+    if Que::Scheduler::ToEnqueue.active_job_sufficient_version?
 
-class SpecifiedByClassTestJob < ::Que::Job
-  def run; end
-end
+      Class.new(::ActiveJob::Base) do
+        self.queue_adapter = :que
 
-class WithArgsTestJob < ::Que::Job
-  def run; end
-end
+        def run; end
+      end
+    else
+      Class.new(::Que::Job) do
+        def run; end
+      end
 
-class DailyTestJob < ::Que::Job
-  def run; end
-end
-
-class TwiceDailyTestJob < ::Que::Job
-  def run; end
-end
-
-class TimezoneTestJob < ::Que::Job
-  def run; end
+    end
+  Object.const_set(name, clazz)
 end
 
 class NotAQueJob