+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
 - General bundle update of test gems [#154](https://github.com/hlascelles/que-scheduler/pull/154)
+- Support ActiveJob [#164](https://github.com/hlascelles/que-scheduler/pull/164)
 
 ## 3.2.8 (2020-02-09)