+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Unreleased
+## 3.3.0 (2020-04-04)
 
 - General bundle update of test gems [#154](https://github.com/hlascelles/que-scheduler/pull/154)
 - Support ActiveJob [#164](https://github.com/hlascelles/que-scheduler/pull/164)